KP-8503 Added initial styling code to match designs
Added font files, logo images, and organized styling files.
Created HSL color variables based on the designs and implemented a way to overwrite the hue and saturation of primary and secondary colors.
Created styles Avatar and Button components.
Implemented styling for login and reset password pages, as well as some initial styling for the header and footer.